Organized loose files
Organized files into folders to make it easier to find algorithms and examples by language.
