+++ C/linkedlist (2).c
